extract CEnvelope into seperate header
